2/4(Tue) - Complete step02 of Mission1 DONE: attempted to diminish the length of methods by spliting previous playGame() method. - The playGame() of step01 was split into two methods: playGame() and raceEachMonster(). FURTHER EXPLANATION: playGame() method is responsible for controlling the order of monsters, while raceEachMonster() method takes the current index from playGame() and iterates the number that user designates before the methods are executed. NOTE: The refactored code passed the test code made in step01.
